Moved resource table rendering logic to separate file
